requirements.txt
azure-iot-device
GitPython